# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для платформы изучения английского языка. Используется Dictionary API для получения данных о словах.",
        "title": "English Learning Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/user-cards": {
            "get": {
                "description": "Возвращает список всех записей прогресса пользователей по карточкам",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-cards"
                ],
                "summary": "Получить все записи прогресса по карточкам",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserCard"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новую запись о прогрессе пользователя по карточке",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-cards"
                ],
                "summary": "Создать запись о прогрессе по карточке",
                "parameters": [
                    {
                        "description": "Данные для создания записи",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateUserCardRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.UserCard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/user-cards/user/{user_id}": {
            "get": {
                "description": "Возвращает список всех карточек конкретного пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-cards"
                ],
                "summary": "Получить карточки пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserCard"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/user-cards/{id}": {
            "get": {
                "description": "Возвращает информацию о прогрессе пользователя по карточке",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-cards"
                ],
                "summary": "Получить прогресс по карточке по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserCard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет информацию о прогрессе пользователя по карточке",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-cards"
                ],
                "summary": "Обновить прогресс по карточке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateUserCardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserCard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет запись о прогрессе пользователя по карточке",
                "tags": [
                    "user-cards"
                ],
                "summary": "Удалить прогресс по карточке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateUserCardRequest": {
            "type": "object",
            "required": [
                "card_id",
                "user_id"
            ],
            "properties": {
                "card_id": {
                    "description": "int64",
                    "type": "integer"
                },
                "user_deck_id": {
                    "description": "int64, опционально",
                    "type": "integer"
                },
                "user_id": {
                    "description": "UUID",
                    "type": "string"
                }
            }
        },
...
        "handlers.UpdateUserCardRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "correct_count": {
                    "type": "integer"
                },
                "last_seen": {
                    "description": "ISO 8601 timestamp",
                    "type": "string"
                },
                "next_review": {
                    "description": "ISO 8601 timestamp",
                    "type": "string"
                },
                "status": {
                    "description": "'new', 'learning', 'learned'",
                    "type": "string"
                },
                "wrong_count": {
                    "type": "integer"
                }
            }
        },
...
        "models.UserCard": {
            "type": "object",
            "properties": {
                "card_id": {
                    "type": "integer"
                },
                "correct_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_seen": {
                    "type": "string"
                },
                "next_review": {
                    "type": "string"
                },
                "status": {
                    "description": "'new', 'learning', 'learned'",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_deck_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "wrong_count": {
                    "type": "integer"
                }
            }
        },
...
    }
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Type Bearer followed by a space and JWT token."
        }
    }
}